@@ -1,10 +1,12 @@
 package model
 
+// TODO shouldn't this rather be TemplateLoader?
+
 // WorkflowLoader is a loader interface required for workflow initialization.
 type WorkflowLoader interface {
-	// ListTags lists a list of templates for tags from the provided templates directory
-	ListTags(workflowTags []string) []string
+	// GetTemplatePathsByTags returns a list of template paths based on the provided tags from the templates directory
+	GetTemplatePathsByTags(tags []string) []string
 
-	// ListTemplates takes a list of templates and returns paths for them
-	ListTemplates(templatesList []string, noValidate bool) []string
+	// GetTemplatePaths takes a list of templates and returns paths for them
+	GetTemplatePaths(templatesList []string, noValidate bool) []string
 }
