@@ -10,18 +10,27 @@ import (
 
 // Operators contains the operators that can be applied on protocols
 type Operators struct {
-	// Matchers contains the detection mechanism for the request to identify
-	// whether the request was successful
+	// description: |
+	//   Matchers contains the detection mechanism for the request to identify
+	//   whether the request was successful by doing pattern matching
+	//   on request/responses.
+	//
+	//   Multiple matchers can be combined together with `matcher-condition` flag
+	//   which accepts either `and` or `or` as argument.
 	Matchers []*matchers.Matcher `yaml:"matchers,omitempty"`
-	// Extractors contains the extraction mechanism for the request to identify
-	// and extract parts of the response.
+	// description: |
+	//   Extractors contains the extraction mechanism for the request to identify
+	//   and extract parts of the response.
 	Extractors []*extractors.Extractor `yaml:"extractors,omitempty"`
-	Sets       []*sets.Set             `yaml:"sets,omitempty"`
-	// MatchersCondition is the condition of the matchers
-	// whether to use AND or OR. Default is OR.
+	// description: |
+	//   MatchersCondition is the condition between the matchers. Default is OR.
+	// values:
+	//   - "and"
+	//   - "or"
 	MatchersCondition string `yaml:"matchers-condition,omitempty"`
 	// cached variables that may be used along with request.
 	matchersCondition matchers.ConditionType
+	Sets              []*sets.Set `yaml:"sets,omitempty"`
 }
 
 // Compile compiles the operators as well as their corresponding matchers and extractors
