@@ -1,13 +1,14 @@
 package parsers
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"os"
+	"regexp"
 
 	"gopkg.in/yaml.v2"
 
+	"github.com/projectdiscovery/gologger"
 	"github.com/projectdiscovery/nuclei/v2/pkg/catalog/loader/filter"
 	"github.com/projectdiscovery/nuclei/v2/pkg/model"
 	"github.com/projectdiscovery/nuclei/v2/pkg/templates"
@@ -16,11 +17,15 @@ import (
 
 const mandatoryFieldMissingTemplate = "mandatory '%s' field is missing"
 
-// Load loads a template by parsing metadata and running all tag and path based filters on the template.
-func Load(templatePath string, isWorkflow bool, workflowTags []string, tagFilter *filter.TagFilter) (bool, error) {
-	template, templateParseError := parseTemplate(templatePath)
+// LoadTemplate returns true if the template is valid and matches the filtering criteria.
+func LoadTemplate(templatePath string, tagFilter *filter.TagFilter, extraTags []string) (bool, error) {
+	template, templateParseError := ParseTemplate(templatePath)
 	if templateParseError != nil {
 		return false, templateParseError
+	}
+
+	if len(template.Workflows) > 0 {
+		return false, nil
 	}
 
 	templateInfo := template.Info
@@ -28,37 +33,41 @@ func Load(templatePath string, isWorkflow bool, workflowTags []string, tagFilter
 		return false, validationError
 	}
 
-	if len(template.Workflows) > 0 {
-		if isWorkflow {
-			return true, nil // if a workflow is declared and this template is a workflow, then load
-		} else { //nolint:indent-error-flow,revive // preferred: readability and extensibility
-			return false, nil // if a workflow is declared and this template is not a workflow then do not load
-		}
-	} else if isWorkflow {
-		return false, nil // if no workflows are declared and this template is a workflow then do not load
-	} else { // if workflows are not declared and the template is not a workflow then parse it
-		return isInfoMetadataMatch(tagFilter, &templateInfo, workflowTags)
-	}
+	return isTemplateInfoMetadataMatch(tagFilter, &templateInfo, extraTags)
 }
 
-func isInfoMetadataMatch(tagFilter *filter.TagFilter, templateInfo *model.Info, workflowTags []string) (bool, error) {
+// LoadWorkflow returns true if the workflow is valid and matches the filtering criteria.
+func LoadWorkflow(templatePath string, tagFilter *filter.TagFilter) (bool, error) {
+	template, templateParseError := ParseTemplate(templatePath)
+	if templateParseError != nil {
+		return false, templateParseError
+	}
+
+	templateInfo := template.Info
+
+	if len(template.Workflows) > 0 {
+		if validationError := validateMandatoryInfoFields(&templateInfo); validationError != nil {
+			return false, validationError
+		}
+
+		return isTemplateInfoMetadataMatch(tagFilter, &templateInfo, nil) // we don't want workflows to be loaded by tags
+	}
+
+	return false, nil
+}
+
+func isTemplateInfoMetadataMatch(tagFilter *filter.TagFilter, templateInfo *model.Info, extraTags []string) (bool, error) {
 	templateTags := templateInfo.Tags.ToSlice()
 	templateAuthors := templateInfo.Authors.ToSlice()
 	templateSeverity := templateInfo.SeverityHolder.Severity
 
-	var match bool
-	var err error
-	if len(workflowTags) == 0 {
-		match, err = tagFilter.Match(templateTags, templateAuthors, templateSeverity)
-	} else {
-		match, err = tagFilter.MatchWithWorkflowTags(templateTags, templateAuthors, templateSeverity, workflowTags)
-	}
+	match, err := tagFilter.Match(templateTags, templateAuthors, templateSeverity, extraTags)
 
 	if err == filter.ErrExcluded {
 		return false, filter.ErrExcluded
 	}
 
-	return match, nil
+	return match, err
 }
 
 func validateMandatoryInfoFields(info *model.Info) error {
@@ -76,7 +85,10 @@ func validateMandatoryInfoFields(info *model.Info) error {
 	return nil
 }
 
-func parseTemplate(templatePath string) (*templates.Template, error) {
+var fieldErrorRegexp = regexp.MustCompile(`not found in`)
+
+// ParseTemplate parses a template and returns a *templates.Template structure
+func ParseTemplate(templatePath string) (*templates.Template, error) {
 	f, err := os.Open(templatePath)
 	if err != nil {
 		return nil, err
@@ -89,8 +101,12 @@ func parseTemplate(templatePath string) (*templates.Template, error) {
 	}
 
 	template := &templates.Template{}
-	err = yaml.NewDecoder(bytes.NewReader(data)).Decode(template)
+	err = yaml.UnmarshalStrict(data, template)
 	if err != nil {
+		if fieldErrorRegexp.MatchString(err.Error()) {
+			gologger.Warning().Msgf("Unrecognized fields in template %s: %s", templatePath, err)
+			return template, nil
+		}
 		return nil, err
 	}
 	return template, nil
