@@ -29,35 +29,33 @@ func NewLoader(options *protocols.ExecuterOptions) (model.WorkflowLoader, error)
 	return &workflowLoader{pathFilter: pathFilter, tagFilter: tagFilter, options: options}, nil
 }
 
-// ListTags lists a list of templates for tags from the provided templates directory
-func (w *workflowLoader) ListTags(workflowTags []string) []string {
+func (w *workflowLoader) GetTemplatePathsByTags(templateTags []string) []string {
 	includedTemplates := w.options.Catalog.GetTemplatesPath([]string{w.options.Options.TemplatesDirectory})
-	templatesMap := w.pathFilter.Match(includedTemplates)
+	templatePathMap := w.pathFilter.Match(includedTemplates)
 
-	loadedTemplates := make([]string, 0, len(templatesMap))
-	for k := range templatesMap {
-		loaded, err := Load(k, false, workflowTags, w.tagFilter)
+	loadedTemplates := make([]string, 0, len(templatePathMap))
+	for templatePath := range templatePathMap {
+		loaded, err := LoadTemplate(templatePath, w.tagFilter, templateTags)
 		if err != nil {
-			gologger.Warning().Msgf("Could not load template %s: %s\n", k, err)
+			gologger.Warning().Msgf("Could not load template %s: %s\n", templatePath, err)
 		} else if loaded {
-			loadedTemplates = append(loadedTemplates, k)
+			loadedTemplates = append(loadedTemplates, templatePath)
 		}
 	}
 	return loadedTemplates
 }
 
-// ListTemplates takes a list of templates and returns paths for them
-func (w *workflowLoader) ListTemplates(templatesList []string, noValidate bool) []string {
+func (w *workflowLoader) GetTemplatePaths(templatesList []string, noValidate bool) []string {
 	includedTemplates := w.options.Catalog.GetTemplatesPath(templatesList)
-	templatesMap := w.pathFilter.Match(includedTemplates)
+	templatesPathMap := w.pathFilter.Match(includedTemplates)
 
-	loadedTemplates := make([]string, 0, len(templatesMap))
-	for k := range templatesMap {
-		matched, err := Load(k, false, nil, w.tagFilter)
+	loadedTemplates := make([]string, 0, len(templatesPathMap))
+	for templatePath := range templatesPathMap {
+		matched, err := LoadTemplate(templatePath, w.tagFilter, nil)
 		if err != nil {
-			gologger.Warning().Msgf("Could not load template %s: %s\n", k, err)
+			gologger.Warning().Msgf("Could not load template %s: %s\n", templatePath, err)
 		} else if matched || noValidate {
-			loadedTemplates = append(loadedTemplates, k)
+			loadedTemplates = append(loadedTemplates, templatePath)
 		}
 	}
 	return loadedTemplates
