@@ -16,39 +16,108 @@ import (
 type Request struct {
 	// Operators for the current request go here.
 	operators.Operators `yaml:",inline"`
-	// Path contains the path/s for the request
-	Path []string `yaml:"path"`
-	// Raw contains raw requests
-	Raw []string `yaml:"raw"`
-	ID  string   `yaml:"id"`
-	// Name is the name of the request
-	Name string `yaml:"Name"`
-	// AttackType is the attack type
-	// Sniper, PitchFork and ClusterBomb. Default is Sniper
-	AttackType string `yaml:"attack"`
-	// Method is the request method, whether GET, POST, PUT, etc
-	Method string `yaml:"method"`
-	// Body is an optional parameter which contains the request body for POST methods, etc
-	Body string `yaml:"body"`
-	// Path contains the path/s for the request variables
-	Payloads map[string]interface{} `yaml:"payloads"`
-	// Headers contains headers to send with the request
-	Headers map[string]string `yaml:"headers"`
-	// RaceNumberRequests is the number of same request to send in race condition attack
-	RaceNumberRequests int `yaml:"race_count"`
-	// MaxRedirects is the maximum number of redirects that should be followed.
-	MaxRedirects int `yaml:"max-redirects"`
-	// PipelineConcurrentConnections is number of connections in pipelining
-	PipelineConcurrentConnections int `yaml:"pipeline-concurrent-connections"`
-	// PipelineRequestsPerConnection is number of requests in pipelining
-	PipelineRequestsPerConnection int `yaml:"pipeline-requests-per-connection"`
-	// Threads specifies number of threads for sending requests
-	Threads int `yaml:"threads"`
+	// description: |
+	//   Path contains the path/s for the HTTP requests. It supports variables
+	//   as placeholders.
+	// examples:
+	//   - name: Some example path values
+	//     value: >
+	//       []string{"{{BaseURL}}", "{{BaseURL}}/+CSCOU+/../+CSCOE+/files/file_list.json?path=/sessions"}
+	Path []string `yaml:"path,omitempty"`
+	// description: |
+	//   Raw contains HTTP Requests in Raw format.
+	// examples:
+	//   - name: Some example raw requests
+	//     value: |
+	//       []string{"GET /etc/passwd HTTP/1.1\nHost:\nContent-Length: 4", "POST /.%0d./.%0d./.%0d./.%0d./bin/sh HTTP/1.1\nHost: {{Hostname}}\nUser-Agent: Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:71.0) Gecko/20100101 Firefox/71.0\nContent-Length: 1\nConnection: close\n\necho\necho\ncat /etc/passwd 2>&1"}
+	Raw []string `yaml:"raw,omitempty"`
+	// ID is the ID of the request
+	ID string `yaml:"id,omitempty"`
+	// description: |
+	//  Name is the optional name of the request.
+	//
+	//  If a name is specified, all the named request in a template can be matched upon
+	//  in a combined manner allowing multirequest based matchers.
+	Name string `yaml:"name,omitempty"`
+	// description: |
+	//   Attack is the type of payload combinations to perform.
+	//
+	//   Sniper is each payload once, pitchfork combines multiple payload sets and clusterbomb generates
+	//   permutations and combinations for all payloads.
+	// values:
+	//   - "sniper"
+	//   - "pitchfork"
+	//   - "clusterbomb"
+	AttackType string `yaml:"attack,omitempty"`
+	// description: |
+	//   Method is the HTTP Request Method.
+	// values:
+	//   - "GET"
+	//   - "POST"
+	//   - "PUT"
+	//   - "DELETE"
+	Method string `yaml:"method,omitempty"`
+	// description: |
+	//   Body is an optional parameter which contains HTTP Request body.
+	// examples:
+	//   - name: Same Body for a Login POST request
+	//     value: "\"username=test&password=test\""
+	Body string `yaml:"body,omitempty"`
+	// description: |
+	//   Payloads contains any payloads for the current request.
+	//
+	//   Payloads support both key-values combinations where a list
+	//   of payloads is provided, or optionally a single file can also
+	//   be provided as payload which will be read on run-time.
+	Payloads map[string]interface{} `yaml:"payloads,omitempty"`
+	// description: |
+	//   Headers contains HTTP Headers to send with the request.
+	// examples:
+	//   - value: |
+	//       map[string]string{"Content-Type": "application/x-www-form-urlencoded", "Content-Length": "1", "Any-Header": "Any-Value"}
+	Headers map[string]string `yaml:"headers,omitempty"`
+	// description: |
+	//   RaceCount is the number of times to send a request in Race Condition Attack.
+	// examples:
+	//   - name: Send a request 5 times
+	//     value: "5"
+	RaceNumberRequests int `yaml:"race_count,omitempty"`
+	// description: |
+	//   MaxRedirects is the maximum number of redirects that should be followed.
+	// examples:
+	//   - name: Follow upto 5 redirects
+	//     value: "5"
+	MaxRedirects int `yaml:"max-redirects,omitempty"`
+	// description: |
+	//   PipelineConcurrentConnections is number of connections to create during pipelining.
+	// examples:
+	//   - name: Create 40 concurrent connections
+	//     value: 40
+	PipelineConcurrentConnections int `yaml:"pipeline-concurrent-connections,omitempty"`
+	// description: |
+	//   PipelineRequestsPerConnection is number of requests to send per connection when pipelining.
+	// examples:
+	//   - name: Send 100 requests per pipeline connection
+	//     value: 100
+	PipelineRequestsPerConnection int `yaml:"pipeline-requests-per-connection,omitempty"`
+	// description: |
+	//   Threads specifies number of threads to use sending requests. This enables Connection Pooling.
+	//
+	//   Connection: Close attribute must not be used in request while using threads flag, otherwise
+	//   pooling will fail and engine will continue to close connections after requests.
+	// examples:
+	//   - name: Send requests using 10 concurrent threads
+	//     value: 10
+	Threads int `yaml:"threads,omitempty"`
 
-	// MaxSize is the maximum size of http response body to read in bytes.
-	MaxSize int `yaml:"max-size"`
+	// description: |
+	//   MaxSize is the maximum size of http response body to read in bytes.
+	// examples:
+	//   - name: Read max 2048 bytes of the response
+	//     value: 2048
+	MaxSize int `yaml:"max-size,omitempty"`
 
-	CompiledOperators *operators.Operators
+	CompiledOperators *operators.Operators `yaml:"-"`
 
 	options       *protocols.ExecuterOptions
 	attackType    generators.Type
@@ -57,22 +126,37 @@ type Request struct {
 	generator     *generators.Generator // optional, only enabled when using payloads
 	httpClient    *retryablehttp.Client
 	rawhttpClient *rawhttp.Client
-	// CookieReuse is an optional setting that makes cookies shared within requests
-	CookieReuse bool `yaml:"cookie-reuse"`
-	// Redirects specifies whether redirects should be followed.
-	Redirects bool `yaml:"redirects"`
-	// Pipeline defines if the attack should be performed with HTTP 1.1 Pipelining (race conditions/billions requests)
-	// All requests must be indempotent (GET/POST)
-	Pipeline bool `yaml:"pipeline"`
-	// Specify in order to skip request RFC normalization
-	Unsafe bool `yaml:"unsafe"`
-	// Race determines if all the request have to be attempted at the same time
-	// The minimum number of requests is determined by threads
-	Race bool `yaml:"race"`
-	// ReqCondition automatically assigns numbers to requests and preserves
-	// their history for being matched at the end.
-	// Currently only works with sequential http requests.
-	ReqCondition bool `yaml:"req-condition"`
+
+	// description: |
+	//   CookieReuse is an optional setting that enables cookie reuse for
+	//   all requests defined in raw section.
+	CookieReuse bool `yaml:"cookie-reuse,omitempty"`
+	// description: |
+	//   Redirects specifies whether redirects should be followed by the HTTP Client.
+	//
+	//   This can be used in conjunction with `max-redirects` to control the HTTP request redirects.
+	Redirects bool `yaml:"redirects,omitempty"`
+	// description: |
+	//   Pipeline defines if the attack should be performed with HTTP 1.1 Pipelining
+	//
+	//   All requests must be indempotent (GET/POST). This can be used for race conditions/billions requests.
+	Pipeline bool `yaml:"pipeline,omitempty"`
+	// description: |
+	//   Unsafe specifies whether to use rawhttp engine for sending Non RFC-Compliant requests.
+	//
+	//   This uses the [rawhttp](https://github.com/projectdiscovery/rawhttp) engine to achieve complete
+	//   control over the request, with no normalization performed by the client.
+	Unsafe bool `yaml:"unsafe,omitempty"`
+	// description: |
+	//   Race determines if all the request have to be attempted at the same time (Race Condition)
+	//
+	//   The actual number of requests that will be sent is determined by the `race_count`  field.
+	Race bool `yaml:"race,omitempty"`
+	// description: |
+	//   ReqCondition automatically assigns numbers to requests and preserves their history.
+	//
+	//   This allows matching on them later for multi-request conditions.
+	ReqCondition bool `yaml:"req-condition,omitempty"`
 }
 
 // GetID returns the unique ID of the request if any.
