@@ -1,3 +1,4 @@
+//go:generate dstdocgen -path "" -structure Template -output templates_doc.go -package templates
 package templates
 
 import (
@@ -12,26 +13,56 @@ import (
 	"github.com/projectdiscovery/nuclei/v2/pkg/workflows"
 )
 
-// Template is a request template parsed from a yaml file
+// Template is a YAML input file which defines all the requests and
+// other metadata for a template.
 type Template struct {
-	// ID is the unique id for the template
+	// description: |
+	//   ID is the unique id for the template. IDs must be lowercase
+	//   and must not contain spaces in it.
+	//
+	//   #### Good IDs
+	//
+	//   A good ID uniquely identifies what the requests in the template
+	//   are doing. Let's say you have a template that identifies a git-config
+	//   file on the webservers, a good name would be `git-config-exposure`. Another
+	//   example name is `azure-apps-nxdomain-takeover`.
+	// examples:
+	//   - name: ID Example
+	//     value: "\"cve-2021-19520\""
 	ID string `yaml:"id"`
-	// Info contains information about the template
+	// description: |
+	//   Info contains metadata information about the template.
+	// examples:
+	//   - value: exampleInfoStructure
 	Info model.Info `yaml:"info"`
-	// RequestsHTTP contains the http request to make in the template
+	// description: |
+	//   Requests contains the http request to make in the template.
+	// examples:
+	//   - value: exampleNormalHTTPRequest
 	RequestsHTTP []*http.Request `yaml:"requests,omitempty" json:"requests"`
-	// RequestsDNS contains the dns request to make in the template
+	// description: |
+	//   DNS contains the dns request to make in the template
+	// examples:
+	//   - value: exampleNormalDNSRequest
 	RequestsDNS []*dns.Request `yaml:"dns,omitempty" json:"dns"`
-	// RequestsFile contains the file request to make in the template
+	// description: |
+	//   File contains the file request to make in the template
+	// examples:
+	//   - value: exampleNormalFileRequest
 	RequestsFile []*file.Request `yaml:"file,omitempty" json:"file"`
-	// RequestsNetwork contains the network request to make in the template
+	// description: |
+	//   Network contains the network request to make in the template
+	// examples:
+	//   - value: exampleNormalNetworkRequest
 	RequestsNetwork []*network.Request `yaml:"network,omitempty" json:"network"`
-	// RequestsHeadless contains the headless request to make in the template.
+	// description: |
+	//   Headless contains the headless request to make in the template.
 	RequestsHeadless []*headless.Request `yaml:"headless,omitempty" json:"headless"`
 	// RequestCode contains the nebula script to be executed
 	RequestCode []*code.Request `yaml:"code,omitempty" json:"code"`
 
-	// Workflows is a yaml based workflow declaration code.
+	// description: |
+	//   Workflows is a yaml based workflow declaration code.
 	workflows.Workflow `yaml:",inline,omitempty"`
 	CompiledWorkflow   *workflows.Workflow `yaml:"-" json:"-" jsonschema:"-"`
 
