@@ -7,7 +7,8 @@ import (
 
 // Workflow is a workflow to execute with chained requests, etc.
 type Workflow struct {
-	// Workflows is a yaml based workflow declaration code.
+	// description: |
+	//   Workflows is a list of workflows to execute for a template.
 	Workflows []*WorkflowTemplate `yaml:"workflows,omitempty"`
 
 	Options *protocols.ExecuterOptions
@@ -15,16 +16,25 @@ type Workflow struct {
 
 // WorkflowTemplate is a template to be ran as part of a workflow
 type WorkflowTemplate struct {
-	// Template is the template to run
-	Template string `yaml:"template"`
-	// Tags to perform filtering of supplied templates on
-	Tags model.StringSlice `yaml:"tags"`
-	// Matchers perform name based matching to run subtemplates for a workflow.
-	Matchers []*Matcher `yaml:"matchers"`
-	// Subtemplates are ran if the template matches.
-	Subtemplates []*WorkflowTemplate `yaml:"subtemplates"`
+	// description: |
+	//   Template is a single template or directory to execute as part of workflow.
+	// examples:
+	//   - name: A single template
+	//     value: "\"dns/worksites-detection.yaml\""
+	//   - name: A template directory
+	//     value: "\"misconfigurations/aem\""
+	Template string `yaml:"template,omitempty"`
+	// description: |
+	//    Tags to run templates based on.
+	Tags model.StringSlice `yaml:"tags,omitempty"`
+	// description: |
+	//    Matchers perform name based matching to run subtemplates for a workflow.
+	Matchers []*Matcher `yaml:"matchers,omitempty"`
+	// description: |
+	//    Subtemplates are ran if the `template` field Template matches.
+	Subtemplates []*WorkflowTemplate `yaml:"subtemplates,omitempty"`
 	// Executers perform the actual execution for the workflow template
-	Executers []*ProtocolExecuterPair
+	Executers []*ProtocolExecuterPair `yaml:"-"`
 }
 
 // ProtocolExecuterPair is a pair of protocol executer and its options
@@ -35,8 +45,10 @@ type ProtocolExecuterPair struct {
 
 // Matcher performs conditional matching on the workflow template results.
 type Matcher struct {
-	// Name is the name of the item to match.
-	Name string `yaml:"name"`
-	// Subtemplates are ran if the name of matcher matches.
-	Subtemplates []*WorkflowTemplate `yaml:"subtemplates"`
+	// description: |
+	//    Name is the name of the item to match.
+	Name string `yaml:"name,omitempty"`
+	// description: |
+	//    Subtemplates are ran if the name of matcher matches.
+	Subtemplates []*WorkflowTemplate `yaml:"subtemplates,omitempty"`
 }
